coopNpcTeam, Auto NPC Team Test Code made operational
This is meant to allow levels NPC Teammates which are added and removed by the coop mod, based on how many players are on the server. The NPC that should be added are stored in a ini file with the name of the level.
If the first player joins the NPC are added, and follow the player.
If more players join these NPC are removed to match a specific number of players/NPC active on the server.
If a player leaves, a NPC is added back in.

This obviously has a long way to go and will require a lot of testing on all viable maps.
The maps will also require changes, how ever, this could be a nice little feature.